Rename _tests to tests
