+++ tests/bot/bot_test.go
