+++ tests/bot/commander_test.go
