+++ tests/bot/config_test.go
