+++ tests/bot/parser_test.go
