+++ tests/bot/rpc_test.go
