+++ tests/template/premade-templates/go_module.txt
