+++ tests/template/premade-templates/js_module.txt
