+++ tests/template/premade-templates/py_module.txt
